requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
pyyaml>=6.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0
tqdm>=4.65.0
tensorboard>=2.14.0
einops>=0.6.1
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
